Added core funtionality for Color Stroop Test.
Scene includes stimulus above 4 interactable cubes. User must select the cube that represents the color of the text presented, not the meaning of the word. Data is saved to JSON and CSV. When running, you can use the VR Simulator to test. Use T and Y to toggle between controllers, and G to select a cube.