Cognitive domain tests grid (#1)
* blank scene

Create blank scene for cognitive-domain-test

* Add shapes and script for shape selection into bins

TODO: fix input system

* Update Interaction calls

Update to use the input actions that are already used in the project, it now works.

* Make New Cognitive Scene

New Cognitive Scene for the grid memorization game. Note: the grid generates outside of the camera, needs to be fixed

* Add grid with pattern generation and selection capability

Grid generates outside of camera, still need to fix this.

* Fix grid positioning and add functionality for game states, fully working without pattern validation

* Implement pattern validation, fix bugs with cube highlight resetting

* Grid Game update for recording

* preparing to merge

---------

Co-authored-by: Rikhil Fellner <rfellner@unc.edu>